Added version 0.01 via upload
This version masks the middle 100 of each sequence of length 300, this proves highly ineffective with the model relying heavily on the edge 100 of either side to predict results. This leads to high amounts of residuals with each consecutive sequence, as due to BPM not remaining constant, the phase gradually shifts.